requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
httpx==0.25.2
twilio==8.10.0
cloudinary==1.36.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-genai>=1.70.0
slowapi==0.1.9
qrcode[pil]==7.4.2
pillow==10.1.0
pytz==2023.3
python-multipart==0.0.6
stripe==7.5.0
